FIX:DataRecordMetadata - record size can be int (not short)
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@3082 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/metadata/DataRecordMetadata.java
@@ -407,18 +407,18 @@ public class DataRecordMetadata implements Serializable {
 	    return recordSize;
 	}
 
-	public short getRecordSizeStripAutoFilling() {
+	public int getRecordSizeStripAutoFilling() {
 		if (recType != FIXEDLEN_RECORD) {
 			return -1;	// unknown size
 		}
 		// compute size of fixed record (without trailing filler)
-		short recSize = (short)0;
-		short prevEnd = 0;
+		int recSize = 0;
+		int prevEnd = 0;
 		for (Object obj: fields) {
 			DataFieldMetadata field = (DataFieldMetadata)obj;
 			if (field.getAutoFilling() == null) {
 				prevEnd += field.getSize() + field.getShift();
-				recSize = (short)Math.max(recSize, prevEnd);
+				recSize = Math.max(recSize, prevEnd);
 			}
 		}
 		return recSize;
